refactor(desktop): extract parseBootstrap to side-effect-free renderer/bootstrap.ts
Move the bootstrap parser and descriptor guards into a side-effect-free renderer module.

Tests can now import the pure parser without loading the DOM entry point.

Kept refreshBrokerStatus in main.ts because it mutates renderer module state.

It also schedules probe side effects and rerenders the status pane.
